"부족한 금액 계산하기" 문제 솔루션2 오류 수정
큰 수가 int 자료형 변수에 저장되어 변형되는 것을 방지하도록 매개변수를 long 자료형으로 변경.

+++ Practice/Programmers/Weekly-Challenge/부족한 금액 계산하기/CalculateTheInsufficientAmount-ArithmeticalProgressionFormula.java
@@ -7,7 +7,7 @@ public class CalculateTheInsufficientAmount {
 	}
 	
 	//부족한 금액 계산하기
-	public static long solution(int price, int money, int count) {
+	public static long solution(long price, long money, long count) {
 		return Math.max((count * (2 * price + (count - 1) * price)) / 2 - money, 0);
 	}
 }
